DEX-1034 respond to code review feedback
